requirements.txt
arabic-reshaper==2.1.3
asgiref==3.5.0
asn1crypto==1.5.1
astroid==2.11.2
autopep8==1.6.0
backports.zoneinfo==0.2.1
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.3
cryptography==37.0.2
cssselect2==0.6.0
dill==0.3.4
dj-database-url==0.5.0
Django==4.0.3
django-entangled==0.5.3
django-on-heroku==1.1.2
django-phonenumber-field==6.1.0
django-widget-tweaks==1.4.12
future==0.18.2
html5lib==1.1
idna==3.3
isort==5.10.1
lazy-object-proxy==1.7.1
lxml==4.8.0
mccabe==0.7.0
oscrypto==1.3.0
phonenumbers==8.12.46
Pillow==9.0.1
pip-autoremove==0.10.0
platformdirs==2.5.1
psycopg2-binary==2.9.3
pycodestyle==2.8.0
pycparser==2.21
pyHanko==0.13.1
pyhanko-certvalidator==0.19.5
pylint==2.13.4
pyotp==2.6.0
PyPDF3==1.0.6
python-bidi==0.4.2
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
qrcode==7.3.1
reportlab==3.6.9
requests==2.27.1
six==1.16.0
sqlparse==0.4.2
svglib==1.2.1
tinycss2==1.1.1
toml==0.10.2
tomli==2.0.1
tqdm==4.64.0
typing-extensions==4.1.1
tzdata==2022.1
tzlocal==4.2
uritools==4.0.0
urllib3==1.26.9
webencodings==0.5.1
whitenoise==6.0.0
wrapt==1.14.0
